Add: ERD변경사항 엔티티 적용 및 liquibase 통한 스키마 생성 스크립트 추가 (#9) 개발 DB 셋팅이 포함됨 liquibase 사용 방법은 노션 참고
