+++ .gitignore
@@ -3,3 +3,4 @@
 /knowlly-core/build/
 /.idea/
 /build/
+/knowlly-core/src/main/resources/application-dev.yml
