+++ knowlly-core/src/main/java/kr/co/knowledgerally/core/lecture/entity/LectureImage.java
@@ -21,11 +21,15 @@ public class LectureImage {
 
     @MapsId("id")
     @ManyToOne
-    @JoinColumn(name = "lecture_id")
-    private Lecture lecture;
+    @JoinColumn(name = "lecture_info_id")
+    private LectureInformation lecture;
 
     @Column(nullable = false)
     private String lectureImgUrl;
+
+    @Builder.Default
+    @Column(nullable = false)
+    private boolean isActive = true;
 
     @Builder.Default
     @CreationTimestamp
