+++ knowlly-core/src/main/java/kr/co/knowledgerally/core/user/entity/Coach.java
@@ -27,6 +27,9 @@ public class Coach {
     @Column(nullable = false)
     private String introduce;
 
+    @Column(nullable = false)
+    private String kakaoId;
+
     @Builder.Default
     @CreationTimestamp
     @Column(nullable = false)
