+++ knowlly-core/src/main/java/kr/co/knowledgerally/core/user/entity/Review.java
@@ -33,6 +33,10 @@ public class Review {
     private String content;
 
     @Builder.Default
+    @Column(nullable = false)
+    private boolean isActive = true;
+
+    @Builder.Default
     @CreationTimestamp
     @Column(nullable = false)
     private LocalDateTime createdAt = LocalDateTime.now();
